pass status as arugument instead of namespacing it
